+++ common/src/main/java/io/github/mortuusars/exposure/Config.java
@@ -56,6 +56,9 @@ public class Config {
         public static final ModConfigSpec.BooleanValue FILM_ROLL_EASY_RENAMING;
         public static final ModConfigSpec.BooleanValue INTERPLANAR_PROJECTOR_LARGER_RENAMING_LIMIT;
 
+        // Integration
+        public static final ModConfigSpec.BooleanValue CREATE_DEPLOYER_STAND_HOTSWAP;
+
         // Debug
         public static final ModConfigSpec.BooleanValue CLEANUP_TIMED_OUT_EXPECTED_EXPOSURES;
 
@@ -189,6 +192,14 @@ public class Config {
                         .comment("Increases item name length limit for Interplanar Projector to 150 characters. Vanilla limit: 50.",
                                 "Default: true")
                         .define("increase_interplanar_projector_name_limit", true);
+            }
+            builder.pop();
+
+            builder.push("integration");
+            {
+                CREATE_DEPLOYER_STAND_HOTSWAP = builder
+                        .comment("Create Deployers will be able to insert/swap attachments on Camera Stand. Default: true")
+                        .define("create_deployer_stand_hotswap", true);
             }
             builder.pop();